fix(build): fix ction conflicts on hot-reload
