+++ packages/brisa/src/utils/response-action/index.test.ts
@@ -9,7 +9,7 @@ import {
 } from 'bun:test';
 import path from 'node:path';
 import extendRequestContext from '@/utils/extend-request-context';
-import responseAction, { getActionFile } from '.';
+import responseAction from '.';
 import { getConstants } from '@/constants';
 import { ENCRYPT_NONTEXT_PREFIX, encrypt } from '@/utils/crypto';
 import { boldLog } from '@/utils/log/log-color';
@@ -845,12 +845,6 @@ describe('utils', () => {
       const res = await responseAction(req);
 
       expect(await res.text()).toBe(JSON.stringify([['foo', 'bar']]));
-    });
-  });
-
-  describe('getActionFile', () => {
-    it('should return the file with the .js extension to work in Node.js and Bun.js', () => {
-      expect(getActionFile('a1_1')).toBe('a1.js');
     });
   });
 });
