+++ packages/brisa/src/utils/transpile-actions/index.ts
@@ -24,35 +24,6 @@ const FN_EXPRESSION_TYPES = new Set([
   'FunctionExpression',
 ]);
 
-export default async function compileActions({
-  actionsEntrypoints,
-  define,
-}: CompileActionsParams) {
-  const { BUILD_DIR, IS_PRODUCTION, CONFIG } = getConstants();
-  const isNode = CONFIG.output === 'node' && IS_PRODUCTION;
-  const rawActionsDir = join(BUILD_DIR, 'actions_raw');
-  const external = CONFIG.external ? [...CONFIG.external, 'brisa'] : ['brisa'];
-  const res = await Bun.build({
-    entrypoints: actionsEntrypoints,
-    outdir: join(BUILD_DIR, 'actions'),
-    external,
-    sourcemap: IS_PRODUCTION ? undefined : 'inline',
-    root: rawActionsDir,
-    target: isNode ? 'node' : 'bun',
-    minify: IS_PRODUCTION,
-    splitting: true,
-    define,
-  });
-
-  if (!res.success) {
-    logBuildError('Failed to compile actions', res.logs);
-  }
-
-  fs.rmSync(rawActionsDir, { recursive: true });
-
-  return res;
-}
-
 export function transpileActions(code: string) {
   let ast = parseCodeToAST(code);
 
@@ -617,4 +588,40 @@ function wrapWithTypeCatch({
       },
     ],
   };
+}
+
+export async function buildActions({
+  actionsEntrypoints,
+  define,
+}: CompileActionsParams) {
+  const { BUILD_DIR, IS_PRODUCTION, CONFIG } = getConstants();
+  const isNode = CONFIG.output === 'node' && IS_PRODUCTION;
+  const rawActionsDir = join(BUILD_DIR, 'actions_raw');
+  const barrelFile = join(rawActionsDir, 'index.ts');
+
+  await Bun.write(
+    barrelFile,
+    actionsEntrypoints.map((p) => `export * from '${p}'`).join('\n'),
+  );
+
+  const external = CONFIG.external ? [...CONFIG.external, 'brisa'] : ['brisa'];
+  const res = await Bun.build({
+    entrypoints: [barrelFile],
+    outdir: join(BUILD_DIR, 'actions'),
+    external,
+    sourcemap: IS_PRODUCTION ? undefined : 'inline',
+    root: rawActionsDir,
+    target: isNode ? 'node' : 'bun',
+    minify: IS_PRODUCTION,
+    splitting: true,
+    define,
+  });
+
+  if (!res.success) {
+    logBuildError('Failed to compile actions', res.logs);
+  }
+
+  fs.rmSync(rawActionsDir, { recursive: true });
+
+  return res;
 }
